chore: move validation functions under validations mod
